created namespace for UdpTimedSender
